refactor: move createInput, fileToBase64 to editorjs/utils
