+++ front/src/utils/editorjs/image-plugin/image.css
@@ -1,3 +1,4 @@
-.embed-tool__content{
+.image-tool__content{
+    margin: 1rem 0;
     width: 100%;
 }
